refactor: review code
