+++ src/layout/footer/footer.scss
@@ -2,33 +2,33 @@
 
 footer {
     background-color: black;
-    height: 130px;
+    height: 200px;
     display: flex;
     justify-content: center;
     align-items: center;
     flex-direction: column;
 
+    @media only screen and (min-width: 1024px) {
+        height: 130px;
+    }
+
     img {
-        height: 20px;
-        width: 60px;
-        margin: 20px 0 20px 0;
+        height: 40px;
+        width: auto;
+        margin: 0 0 20px 0;
+
+        @media only screen and (min-width: 1024px) {
+            height: 20px;
+            width: 60px;
+            margin-top: 20px;
+        }
     }
 
     p {
-        font-family: $font-stack;
-    }
+        font-size: 12px;
 
-    @media only screen and (max-width : 375px) {
-        height: 200px;
-
-        img {
-            height: 40px;
-            width: auto;
-            margin-top: 0;
-        }
-
-        p {
-            font-size: 12px;
+        @media only screen and (min-width: 1024px) {
+            font-family: $font-stack;
         }
     }
 }
