+++ src/page/home/home.scss
@@ -1,28 +1,28 @@
 main {
-    padding: 20px 50px 0 50px;
-    height: 100%;
-    width: 100%;
+    padding: 0 20px;
 
-    section {
-        background-color: #F6F6F6;
-        padding: 50px;
-        border-radius: 25px;
-        margin: 30px 0;
-        display: grid;
-        grid-gap: 50px 60px;
-        grid-template-columns: repeat(3, 1fr);
-        grid-template-rows: repeat(7, 340px)
+    @media only screen and (min-width: 1024px) {
+        height: 100%;
+        width: 100%;
+        padding: 20px 50px 0 50px;
     }
 
-    @media only screen and (max-width: 375px) {
-        padding: 0 20px 0 20px;
+    section {
+        padding: 0;
+        background-color: white;
+        display: grid;
+        grid-template-columns: 1fr;
+        grid-template-rows: repeat(20, 255px);
+        grid-gap: 20px;
+        border-radius: 25px;
+        margin: 30px 0;
 
-        section {
-            padding: 0;
-            background-color: white;
-            grid-template-columns: 1fr;
-            grid-template-rows: repeat(6, 255px);
-            grid-gap: 20px;
+        @media only screen and (min-width: 1024px) {
+            padding: 50px;
+            background-color: #F6F6F6;
+            grid-gap: 50px 60px;
+            grid-template-columns: repeat(3, 1fr);
+            grid-template-rows: repeat(7, 340px);
         }
     }
 }
